Remove hardcoded ruby versions.
add param to all tasks that have a hardcoded ruby version to be able to
set them for pipelines that rely on these tasks.

currently if the remote pipeline overrides the image for the task (e.g.
bosh-azure-cpi and bosh-aws-cpi) the chosen ruby version depends on the
availability of chruby. E.g. the bosh-aws-cpi image does not have chruby
and therefore uses the image system ruby (currently 3.1.2). the
bosh-azure-cpi image has chruby available and tries to run with 2.4.4
which is not available in the image.

make the ruby version that the script uses configurable to solve this.
pin default to old hardcoded version to avoid pipelines breaking because
of this change.